move DbTool back to bin/reth
the rest will be needed for node-builder move
