+++ bin/reth/Cargo.toml
@@ -109,6 +109,7 @@ pretty_assertions = "1.3.0"
 humantime = "2.1.0"
 itertools.workspace = true
 rayon.workspace = true
+boyer-moore-magiclen = "0.2.16"
 
 [target.'cfg(not(windows))'.dependencies]
 jemallocator = { version = "0.5.0", optional = true }
