+++ bin/reth/src/lib.rs
@@ -31,6 +31,7 @@ pub mod cli;
 pub mod commands;
 pub mod prometheus_exporter;
 pub mod runner;
+pub mod utils;
 
 /// Re-exported payload related types
 pub mod payload {
@@ -56,12 +57,6 @@ pub mod args {
 /// the `reth_node_core::args` re-export for more details.
 pub mod version {
     pub use reth_node_core::version::*;
-}
-
-/// Re-exported from `reth_node_core`, also to prevent a breaking change. See the comment on
-/// the `reth_node_core::args` re-export for more details.
-pub mod utils {
-    pub use reth_node_core::utils::*;
 }
 
 /// Re-exported from `reth_node_core`, also to prevent a breaking change. See the comment on
